item logic refactoring
